% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_calc.R
\name{calc_performance_metrics}
\alias{calc_performance_metrics}
\title{Calculate diagnostic performance metrics}
\usage{
calc_performance_metrics(
  actual,
  predicted,
  score = NULL,
  metrics_to_calc = c("sensitivity", "specificity", "ppv", "npv", "accuracy", "auc"),
  ci_method_binary = "cp",
  ci_method_auc = "delong",
  alpha = 0.05,
  boot_n = 2000
)
}
\arguments{
\item{actual}{Vectors of actual values(0=negative, 1=positive)}

\item{predicted}{Vector of predicted binary labels}

\item{score}{Optional probability scores for AUC}

\item{metrics_to_calc}{Character vector of metrics to calculate}

\item{ci_method_binary}{Select method to calculate CI for sn,sp,ppv,npv,acc}

\item{ci_method_auc}{Select method to calculate CI for AUC}

\item{alpha}{type I error}

\item{boot_n}{The number of Bootstrap sampling}
}
\value{
A named list of calculated performance metrics
}
\description{
Calculate diagnostic performance metrics
}
